tegra: fb: HDMI monitor spec. not correct via "mode" device.
When monitor is auto-detected on HDMI port, a special structure
is initialised to hold the video mode properties for the plugged-
in monitor. From this the display string is generated for displa
via the "mode" character device under /sys.

When the monitor is disconnected, the structures associated with
the monitor tself are cleared, but the pointer to the video mode
properties structure was left initialised and so pointed into thin
air, which resulted in garbage appearing on the "mode" device.

Added explicit deinitialisation code to reset pointer to video
mode properties so "cat mode" within /sys yields an empty string
when the HDMI monitor is unplugged.

Bug 849877

Reviewed-on: http://git-master/r/41205
(cherry picked from commit 8cecf2d7fbdf4d00945f3dd13bfe00161d4231d5)

Change-Id: Ibf3f070fa7904e8a01966d2886bbe07e9b69a607
Reviewed-on: http://git-master/r/68588
Reviewed-by: Lokesh Pathak <lpathak@nvidia.com>
Tested-by: Lokesh Pathak <lpathak@nvidia.com>

Rebase-Id: R6aaaf42f30258c6537adabf9050e7c37ad7a1807