DTV: dma: fixed dma burst size issue DTV xmit
Since dtv interface was designed based on SPI bus, it shares the same
dma configuration with SPI bus. However, it is not proper because DTV
interface has to transmit data in 4 words long constantly. The patch
added an option in tegra_dma_req to set burst size to be fixed 4 words.

fixed Bug 910227

Change-Id: I1436f0c8d108dd39edc57ae4c9cb750d9574b62c
Reviewed-on: http://git-master/r/75509
Reviewed-by: David Schalig <dschalig@nvidia.com>
Reviewed-by: Laxman Dewangan <ldewangan@nvidia.com>
Signed-off-by: Adam Jiang <chaoj@nvidia.com>
Signed-off-by: Varun Wadekar <vwadekar@nvidia.com>
Reviewed-on: http://git-master/r/76461
Reviewed-by: Automatic_Commit_Validation_User

Rebase-Id: Ra897bc508edda5856784321b7609e3e8fb6a7500