ARM: tegra: add config option for io_dpd
Adding a temporary config option for allowing devices listed in
tegra_list_io_dpd[] to go into DPD state.
Since the pins under sdhci DPD groups are used for different
purposes on different platforms, this list should be either in
platform files or should be guarded by a config option. This change
does the latter for now, until a decision is made on this.

Bug 1036567.
Bug 1040511.

Change-Id: Ic2de96d9e506463f140d7a4e998c7a657c9593bc
Signed-off-by: Vishal Singh <vissingh@nvidia.com>
Reviewed-on: http://git-master/r/141251
Reviewed-by: Mrutyunjay Sawant <msawant@nvidia.com>
Tested-by: Mrutyunjay Sawant <msawant@nvidia.com>

Rebase-Id: R07835d6a216fba7c61b4431818ef72b913694889