video: tegra: host: Simplify register read
Implement 3D/MPE register read via nvhost_job. We create a job that
contains one GATHER with commands to do the register read, and submit
that job via nvhost_channel_submit.

This way it uses the same context switch logic as the normal submit.

To do this, hardware context needs access to the client's memory
manager, and we need to be able to convert a memory handle into an
id. We also need to serialize the register read to happen after last
operation, so we add a flag in job for that.

Change-Id: Ia9f3cc264657e1127ee47facd6920211a2014177
Signed-off-by: Terje Bergstrom <tbergstrom@nvidia.com>
Reviewed-on: http://git-master/r/189566
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Juha Tukkinen <jtukkinen@nvidia.com>
Reviewed-by: Ken Adams <kadams@nvidia.com>