regulator: max77660: Pass regulator platform data pointers in array
In place of passing the limited regulator platform data and number
of platform data, pass platform data of all regulators in array. The
regulator id will direct relate to index of array. If any regulator is
unused in the given platform then NULL can be passed in place of valid
pointer.

Change-Id: Ib9f896e0b2f71c120da3a0348c00bfd8eb41862e
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-on: http://git-master/r/196734