ARM: Tegra: Add power domain for MC clock
This domain is used to monitor the state of memory clients.
When all clients are runtime suspended this domain will
get turned off. If any client wakes up from runtime suspend
state then this domain will be turned on. When the domains is
off, MC clock can be stopped during cpuidle.

Bug 1010971

Change-Id: I608007a0c70ce9c3c18767b76c1cb489c6c3569b
Signed-off-by: Prashant Gaikwad <pgaikwad@nvidia.com>
Reviewed-on: http://git-master/r/197351
Reviewed-by: Simone Willett <swillett@nvidia.com>
Tested-by: Simone Willett <swillett@nvidia.com>