video: tegra: host: Debug dump to use trylock
Use mutex_trylock() instead of mutex_lock() in debug dump. Debug dump
is not safety critical, so it's ok to sometimes dump state of a
channel that has been torn down.

This fixes a deadlock when channel close waits on queue to clear up
with reflock acquired, submit in queue times out, and timeout code
tries to acquire reflock at debug dump.

Change-Id: Ice9bd3c617808d0b3178912358d7c90f30765be4
Signed-off-by: Terje Bergstrom <tbergstrom@nvidia.com>
Reviewed-on: http://git-master/r/201974
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Yury Gribov <ygribov@nvidia.com>
Tested-by: Yury Gribov <ygribov@nvidia.com>
Reviewed-by: Mayuresh Kulkarni <mkulkarni@nvidia.com>
Reviewed-by: Juha Tukkinen <jtukkinen@nvidia.com>
Reviewed-by: Arto Merilainen <amerilainen@nvidia.com>