watchdog: max77660: provide mechanism for resetting WDT in kthread
The Maxim77660 system watchdog timer generates interrupt 2 second before
it expire. Within this 2 second, it need to reset the timer. Sometime, on
cpu stress, it is not possible to reset wdt in 2 sec and in this case
it is require to have more time.

Add high priority kthread which run on more faster way and reset wdt in
much advance of expire. The time offset is configurable through platform
data.

Change-Id: I7610bc301889c0e9397868c990594ee7374948f0
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-on: http://git-master/r/212999
Reviewed-by: Bitan Biswas <bbiswas@nvidia.com>