ARM: tegra: bbc: call pm_wakeup_event() in IRQ if going to suspend
If an IRQ is fired as we are going to suspend then we want to abort
and make sure voltage and frequency remain high enough for BB to
operate.

The suspend function must not ack the IRQ as the handler shall be
fired and do it.

Bug 1294872

Change-Id: I20e3872b58814402e5ee5945bb6d91c2651eaacb
Signed-off-by: Hervé Fache <hfache@nvidia.com>
Reviewed-on: http://git-master/r/239765
(cherry picked from commit 022b42d27f9dcd3614cd44d297326c8778c66840)
Reviewed-on: http://git-master/r/256176
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>