dma-debug: do work only after initialized
DMA actions can happen before the dma-debug api has been initialized,
and because it's not initialized, there is no memory for the entries and
the system disables itself before even having added the debugfs nodes
that could be used for enabling it again. Use another flag for testing
if the debugging utility has been initialized.

Bug 1303110
Bug 1173494

Change-Id: I27c972ede0a1b64caee62c6552f3ea15e21030c3
Signed-off-by: Konsta Holtta <kholtta@nvidia.com>
Reviewed-on: http://git-master/r/245338
Reviewed-by: Hiroshi Doyu <hdoyu@nvidia.com>
Tested-by: Hiroshi Doyu <hdoyu@nvidia.com>