video: tegra: host: Add ISP & VI power gating
ISP and VI are both in VE domain, so they have to have a common
power domain. nvhost_module_add_domain() will not first try to add
to an existing domain with the same name, and if that doesn't work,
creates a new one.

Use this mechanism to put VI and ISP devices in the common VE domain.

Conflicts:
	drivers/video/tegra/host/isp/isp.c
	drivers/video/tegra/host/nvhost_acm.c

Change-Id: Ib630a81ed514c88d9342579655ed80540c7b3c9d
Signed-off-by: Terje Bergstrom <tbergstrom@nvidia.com>
Reviewed-on: http://git-master/r/241905
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>
Tested-by: Bharat Nihalani <bnihalani@nvidia.com>