staging: nvshm: add support for upcoming modem version
A new version of the modem is coming which is backwards compatible
with the previous one (V3). Changing the kernel to accept both the old
and new versions will allow for a smooth transition as nothing opposes
it.

Note: this version has been set to 0x00020001 so that a future change
can restrict the version check to the 'major' part of it, and allow
minor changes to only change the 'minor' part of the version number.

Bug 1273958
Bug 1342556

Signed-off-by: Hervé Fache <hfache@nvidia.com>
Change-Id: I61a144f935404d8ab4562b3dceb4e95f7d1166f9
Reviewed-on: http://git-master/r/247081
(cherry picked from commit 7b8f253dc3363c3fb6e6e6311909b2954237464d)
Signed-off-by: Ajay Nandakumar <anandakumarm@nvidia.com>
Reviewed-on: http://git-master/r/257860
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Dan Willemsen <dwillemsen@nvidia.com>