video: tegra: host: gk20a: Use Hz in clk iface
gk20a clock interface used PLL frequency in MHz in set_rate(),
get_rate() and round_rate(). The common way to deliver clock
frequencies is to give the frequency in Hz for the unit directly.
This patch makes the necessary changes to the clock code to allow
more consistent usage of gk20a clock.

Bug 1340657
Bug 1330780

Change-Id: I32edbf4da2b85fd7808d9e44876f03156306b3e7
Signed-off-by: Arto Merilainen <amerilainen@nvidia.com>
Reviewed-on: http://git-master/r/263233
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Terje Bergstrom <tbergstrom@nvidia.com>