video: tegra: host: gk20a: Fix recover semantics
gk20a_fifo_recover() function had a semantics in which we were
recovering channels. This semantics is wrong as actually the engines
get jammed and they need to be recovered. The fact that this leads
to tearing down the current context (channel) in the engine is simply
a side effect.

This patch reworks the recover function to account following issues:
- semantics is now fixed
- multiple engines can be recovered with a single call
- mmu id is determined correctly for each engine
- the failing channel is determined directly from the status register
  of failing engine

Bug 1351268

Change-Id: I950c3e92f2acccdadf45a9fd2af4efb15c785e38
Signed-off-by: Arto Merilainen <amerilainen@nvidia.com>
Reviewed-on: http://git-master/r/272022
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Terje Bergstrom <tbergstrom@nvidia.com>