devfreq: Round frequencies in transition stats
Transition statistics required that the new frequency was *exactly*
as it was in the table. However, we may unexpectedly select a
frequency that is not in the table (i.e. thermal limit forces us to
select a bit lower frequency, the original DVFS table does not hold
for the current PLL for some reason, etc.)

This patch modifies transition table to select the best candidate
from the frequency table.

Bug 1366760

Change-Id: I4329f77e733f5f1d39b9d9c3aebe3681a0760d90
Reviewed-on: http://git-master/r/273596
Signed-off-by: Arto Merilainen <amerilainen@nvidia.com>
Signed-off-by: Shridhar Rasal <srasal@nvidia.com>
Reviewed-on: http://git-master/r/1160004
Reviewed-by: Automatic_Commit_Validation_User