video: tegra: host: gk20a: Do not suspend if active
gk20a prepare_poweroff callback is called before the userspace is
freezed. As the callback is called beforce userspace is freezed,
the userspace may still do a single submit causing gk20a to power
on again. However, this itself do not prevent suspend and therefore
we may end up to have wrong book keeping about the state of gk20a.

This patch adds a separate callback into gk20a driver that will
make certain that gk20a is turned off before suspend. If the device
is on at this point, we cancel suspend.

Bug 1357150

Change-Id: Ib85a96b6598de52749090ba2356a14562a77a07c
Signed-off-by: Arto Merilainen <amerilainen@nvidia.com>
Reviewed-on: http://git-master/r/275689
Tested-by: Sang-Hun Lee <sanlee@nvidia.com>
Reviewed-by: Kevin Huang (Eng-SW) <kevinh@nvidia.com>
Reviewed-by: Mitch Luban <mluban@nvidia.com>