Revert "Revert "clocksource: pass DT node pointer to init functions""
This reverts commit 06139bf5f233c03ce1aad929e0adf75fc5fb5898.

In cases where we have multiple nodes of the same type, we may need
the node pointer to know which node was matched. Passing the node
pointer also keeps the init function from having to match the node
a 2nd time.

Update bcm2835, vt8500, and tegra20 init functions for the new
function prototype. Further tegra20 clean-ups are in follow-up commit.

Bug 1379817

Change-Id: Icb37a7bbe4e8f673ce1bcd4ecc92de52d12195eb
Signed-off-by: Ajay Nandakumar <anandakumarm@nvidia.com>
Reviewed-on: http://git-master/r/299022
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>
Tested-by: Bharat Nihalani <bnihalani@nvidia.com>