iommu/tegra: smmu: update asid number for tegra210
The number of asid supported by tegra210 is the same as tegra124,
so make corresponding update.

In addition, update the swgids for tegra210. tegra210 support more
than 64 swgids, for now only enable the first 64.

The num_translation_enable, num_asid_security, and ptc_cache_size
are copied from tegra124 entries. This may need to be confirmed for
tegra210.

Change-Id: I5be1336e392e7cd5b7ef7798b340761673173459
Signed-off-by: Bo Yan <byan@nvidia.com>
Reviewed-on: http://git-master/r/281486
Reviewed-by: Automatic_Commit_Validation_User
Tested-by: Alex Waterman <alexw@nvidia.com>
Reviewed-by: Alex Waterman <alexw@nvidia.com>