video: tegra: host: add power domain for host1x
If we enter LP0 from cpuidle all the context of host1x will
be reset as VDD_CORE is turned off in LP0.

Use generic power domain to save and restore the context of
host1x. If host1x is runtime suspended for 2s then it will
get suspended and context is saved. If some request or remote
weakeup is received for it then it is restored.

Bug 1384396

Change-Id: I6313ee49287379820b4f936c0d7cb8433934ce2c
Signed-off-by: Prashant Gaikwad <pgaikwad@nvidia.com>
Signed-off-by: Shridhar Rasal <srasal@nvidia.com>
Reviewed-on: http://git-master/r/299959
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Terje Bergstrom <tbergstrom@nvidia.com>