ARM: tegra: mcerr: Update client arrays
Use a script to generate the client arrays. Also modifies the
usge of the name field in the client and adds a swgid field.
Before 'name' was really the swgid, but now name holds to true
client name instead of the swgid. This allow for more detail
in the error messages printed by mcerr.

Change-Id: I2a7a912524a1d06b121d9567d32d29bc0b5a1282
Signed-off-by: Alex Waterman <alexw@nvidia.com>
Reviewed-on: http://git-master/r/288592
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Krishna Reddy <vdumpa@nvidia.com>