media: nvavp: uninit nvavp in runtime PM
If we enter LP0 from cpuidle all the context of nvavp will
be reset as VDD_CORE is turned off in LP0.

Save and restore the context of nvavp in runtime suspend and
resume. If nvavp is idle for 2s then it will get suspended
and context is saved. If some request or remote weakeup is
received for it then it is restored.

Also, add current task checks to avoid deadlocks in reentrant
calls.

Task1
nvavp_open
  nvavp_init
    pm_runtime_get_sync
      runtime_resume
          nvavp_init (DEADLOCK)

Bug 1254633

Change-Id: I7c3690624c371c2c111ad3d6f4b41b41d5e7aa97
Signed-off-by: Prashant Gaikwad <pgaikwad@nvidia.com>
Reviewed-on: http://git-master/r/304210
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>