video: tegra: host: gk20a: PMU survives recovery
Engine recovery used to tear down PMU and then restore its state.
This was required as PMU initialisation had steps that affected GR
and FECS. Currently, that part of initialisation is separated from
the bootup part and hence we can also simplify engine recovery.

This patch modifies the recovery routine so that we
1) disable only ELPG
2) recover engines
3) do second stage init for PMU (this includes GR and FECS parts)
4) Re-enable ELPG

Change-Id: I5e3409a7cecbc38b2e43cb11410592ddc9dbc501
Signed-off-by: Arto Merilainen <amerilainen@nvidia.com>
Reviewed-on: http://git-master/r/327163