video: tegra: host: podgov: Don't turn on the dev
This far we have turned on the device in all code paths that
may lead to scaling. In most cases this is correct, however,
near suspend we may end up powering up the device unnecessarily.

This patch modifies podgov to not turn on device if it is not
already powered. If the device is powered, we take a reference to
ensure that the device stays on over the operations.

In addition, this patch revisits all cancel(_delayed)_work(_sync)
users to make sure that the workers are actually finished.

Bug 1404859

Change-Id: I40608420d0d87aea5aa73fa601b5c202025923e2
Signed-off-by: Arto Merilainen <amerilainen@nvidia.com>