EDP: add temperature dependant rbat table support
Previously, only state of charge was considered when determining
battery impedance. In practice, impedance is dependant on temperature
and state of charge.

This patch adds support for two dimensional battery impedance
(rbat_lut) tables. The rbat value is bi-linearly interpolated from
table as a function of temperature and state of charge.

bug 1412781

Change-Id: I8656aa13f0ebebb848f536194b37420a0438d2ac
Signed-off-by: Timo Alho <talho@nvidia.com>
Reviewed-on: http://git-master/r/335070
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Juha Tukkinen <jtukkinen@nvidia.com>