video: tegra: host: gk20a: update ELPG debugfs
Separate out the elpg transitions counter to an individual
debugfs node. This will make it easier to parse, when read.

Also improve the ELPG debugfs nodes to not unconditionally wake the
GPU up while polling for values. Instead, it should only ensure it
remains powered on if it is already on, and otherwise just return the
values cached at the last railgate entry.

Bug 1410515
Bug 1408385

Change-Id: Id9813737f11942ec8d74075e9cb6fc7c4a50b404
Signed-off-by: Prashant Malani <pmalani@nvidia.com>
Reviewed-on: http://git-master/r/338061
Reviewed-by: Terje Bergstrom <tbergstrom@nvidia.com>
Tested-by: Terje Bergstrom <tbergstrom@nvidia.com>