video: tegra: host: gk20a: Fix va allocation
VA allocation had a bug when no preferred begin address was specified.
The bug was caused by the changed nvhost allocator API which used to
take ~0 address to indicate that the allocator should take any memory
address *but* it currently uses 0 to indicate that no preferred va
is available.

Change-Id: I8a92c65f333eae0402587da19c75df6b91beab64
Signed-off-by: Arto Merilainen <amerilainen@nvidia.com>
Reviewed-on: http://git-master/r/338244
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Terje Bergstrom <tbergstrom@nvidia.com>