video: tegra: dc: change pr macros
It is possible not to have cmu, smart dimmer,
HDMI display timings and it does not mean
problems or warning.
Having tmds node, but no configuration is not
harmful, either. For these, change pr_warn
to pr_info.
dc-default-out node is mandatory. It must be
jumped into failure case if there is not.
For DSI out type, dsi node is mandatory. It must
be jumped into failure case if there is not.

Bug 1371533
Signed-off-by: Min-wuk Lee <mlee@nvidia.com>

Change-Id: Id03f78b293b807c627e83eacf1d99d4922a5d69e
Reviewed-on: http://git-master/r/355070
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Min-wuk Lee <mlee@nvidia.com>
Tested-by: Min-wuk Lee <mlee@nvidia.com>
Reviewed-by: Jon Mayo <jmayo@nvidia.com>