PM/Domain: fix delayed power-off race
Commit 5ffea6f ["CHROMIUM: PM/domains: add delayed power off capability"]
introduced the ability to defer the power-off of a power domain and
used a PM notifier to flush any delayed power-off work for a power
domain. The PM notifiers, however, run before tasks are frozen and
outstanding runtime PM requests are flushed.  A runtime resume may
run during this time scheduling a delayed power-off which may not run
until after pm_genpd_prepare(), where the state of the domain at suspend
is recorded. As a result, a device which expected to be powered off
across suspend may be powered up unexpectedly during resume because
pm_genpd_prepare() recorded the state of the device as on at suspend.
Fix this race by flushing delayed power-off work in pm_genpd_prepare()
instead.

Change-Id: Ia3f83304eaa5eb8a6396175eaa98192a392692ad
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Signed-off-by: Mark Zhang <markz@nvidia.com>
Reviewed-on: http://git-master/r/355390
Reviewed-by: Shridhar Rasal <srasal@nvidia.com>
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>
Tested-by: Bharat Nihalani <bnihalani@nvidia.com>