i2c: mux: pca954x: make pullup regulator optional
The PCA954X driver require two supply, one for self as vcc and
other for output bus pullup as vcc-pullup.

Making vcc-pullup as optional as this is not mandatory regulator
for device and if it is not available then just do not enable it
and not cause to driver fail.

Change-Id: I1c63a13d01ac609af0a74934742daca7e68b4d00
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-on: http://git-master/r/356050
Reviewed-by: Jihoon Bang <jbang@nvidia.com>