pinctrl: tegra: fix drive groups configuration handling
For Tegra124, all drive groups does not supports all
configuration. The configuration which is not supported
is having the bits as -ve in their pinconfig table but
related variable is declared as unsigned making this
configuration as valid.

Correct the datatype of the drive groups configuration bits
and handling this properly across the driver to ignore the
configuration which is not supported.

Change-Id: I9b5666f7fa966df05df566196d155b516cded629
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-on: http://git-master/r/361999