regulator: core: set min/max constraint with configured voltage
If min/max voltage of rail is not provided through constraints
then set this with the already configured voltage.

The motivation is that, if PMIC default is require rail voltage
and do not want to change this voltage then it is not require to
pass the min/max through constraints to avoid mis-configuration
of rail voltage. In this case, update the min/max voltage on
constraints by reading the rail voltage from device directly.

Change-Id: I82befa7ee1fc466569512ad4382246444d0bbb29
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-on: http://git-master/r/362009