video: tegra: host: Inform EDP in prescale
This far EDP was informed about changed load as part of postscale
callback when we changed the frequency. The theory behind this idea
has been that we need to inform EDP due to changed constraint.

However, we have seen cases where updating load information for EDP
would be beneficial more often. This patch modifies the call sequence
so that we inform EDP each time we submit work to GPU.

Bug 1441874

Change-Id: I07a804e0a07ea7efc6024a273437c33a02a8939d
Signed-off-by: Arto Merilainen <amerilainen@nvidia.com>
Reviewed-on: http://git-master/r/362631
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Terje Bergstrom <tbergstrom@nvidia.com>