video: tegra: host: gk20a: Use NV_KIND_DEFAULT
Currently we do not use kind parameter when mapping buffers, however,
we want to start supporting this. We have used 0 as a dummy parameter
in different places in the driver but 0 is actually a valid data type
and hence it cannot be used for identifying "default type". This
patch modifies the calls to use NV_KIND_DEFAULT in these cases.

Bug 1445482

Change-Id: Ib41237e2985746bba669632dee4ab76c12ef036a
Signed-off-by: Arto Merilainen <amerilainen@nvidia.com>
Reviewed-on: http://git-master/r/363304
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Terje Bergstrom <tbergstrom@nvidia.com>