arm: tegra: soctherm: fix PDIV corruption
In the change to adjust on vdd_core vmin, a wrong array index check in
the 'for' loop caused PDIV register to be overwritten with incorrect
values whenever the adjust routine was called. This caused negative
temperatures to show in PLL-therm zone. This also impacts CPU and GPU
temperatures when the rail is below the minimum.

Bug 1363113
Bug 1457777

Change-Id: I67aece5b19821345dcb46e933fabbfe29e98e2cd
Signed-off-by: Diwakar Tundlam <dtundlam@nvidia.com>
Reviewed-on: http://git-master/r/366841
Reviewed-by: Automatic_Commit_Validation_User