video: tegra: host: gk20a: Do not defer ELPG init
Currently the second phase of PMU initialisation runs asynchronously
with other code. This optimises the boot-up time, however, it may
cause unexpected race in PMU code as some functions are called while
the ELPG is still being initialised.

This patch modifies the ELPG initialisation to happen synchronously
as part of power on sequence.

Bug 1456985
Bug 1460116
Bug 1458353

Change-Id: Ic979b3e054a17c0bc6979da2cf03d98f8d416c26
Signed-off-by: Arto Merilainen <amerilainen@nvidia.com>
Reviewed-on: http://git-master/r/368702
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>
Tested-by: Bharat Nihalani <bnihalani@nvidia.com>