arm: tegra: soctherm: move PMC register accesses into PMC driver
The PMC IP block is not part of the SOC_THERM IP block, but the
SOC_THERM code contains direct register reads and writes to the PMC IP
block.  Move these accesses into something PMC-specific: in this case,
drivers/platform/tegra/pmc.c.

While here, remove the usage of the REG_SET/GET* macros, since
upstream Linux practice is to use the actual bit-manipulation
operations, rather than these macros.

This is part of the process of modifying the tegra11_soctherm.c code
to convert it into a low-level device driver for the arsoc_therm IP
block.

We also relocate struct tegra_tsensor_pmu_data to a PMC-specific header
file, since it is PMC-specific, and convert any header references to the
old location to point to the new location.

Thanks to Matt Longnecker for comments on the first version of this patch.

Bug 1201644
Bug 1380438
Bug 1482001

Change-Id: Iea630ac9d9b3dfaab03edf44e2a2725174c7a3d8
Signed-off-by: Paul Walmsley <pwalmsley@nvidia.com>
Cc: Diwakar Tundlam <dtundlam@nvidia.com>
Cc: Matthew Longnecker <mlongnecker@nvidia.com>
Reviewed-on: http://git-master/r/392198
Reviewed-by: Diwakar Tundlam <dtundlam@nvidia.com>
Tested-by: Diwakar Tundlam <dtundlam@nvidia.com>