arm: tegra: soctherm: align usage of PMC_SENSOR_CTRL.BLOCK_SCRATCH_WRITE with the spec
The SOC_THERM code attempts to write a '1' to
PMC_SENSOR_CTRL.BLOCK_SCRATCH_WRITE to attempt to allow it to write to
PMC scratch registers.  However, a value of '1' indicates that scratch
register writes should be *blocked*, not allowed.  If this write were
allowed to succeed - which it is not by the IP block - it would mean
that the subsequent writes in the SOC_THERM code to program the PMIC
shutdown commands into the PMC scratch registers are ignored.  This
would prevent the boot ROM thermtrip code from instructing the PMIC to
shut down after a thermtrip reset initiated by SOC_THERM.

Align the code with the PMC specification by explicitly clearing this
bit.

This patch currently has no practical impact since the boot ROM code
currently clears this bit on at least some Tegra SoCs.  However, by
the specification, the kernel should take care to clear it to ensure
that it can write to the PMC scratch registers.

Of course, the SOC_THERM code should not be directly accessing
PMC registers, since they are in a different IP block.

Thanks to Aleks Frid for some discussion on this patch.

Bug 1482109

Change-Id: I16d78e37143f6d63f2e772718bd8d7ff9a2d9f09
Signed-off-by: Paul Walmsley <pwalmsley@nvidia.com>
Cc: Diwakar Tundlam <dtundlam@nvidia.com>
Cc: Aleksandr Frid <afrid@nvidia.com>
Reviewed-on: http://git-master/r/382272
Reviewed-by: Diwakar Tundlam <dtundlam@nvidia.com>
Tested-by: Diwakar Tundlam <dtundlam@nvidia.com>