arm: tegra: soctherm: check HW max and min temp
Soctherm trip points are set in HW registers that take temperature in
the range -127C to +127C. This causes the bug# noted below.

Fix by checking temp before changing HW register. If threshold temp is
over the range HW can support, we set it to the max.

Also added a change to show the TSOSC STOP bit in debug output.

Bug 1480200

Change-Id: I3c69ca1978ad17ec19af3e896d7e537fb5986ea5
Signed-off-by: Diwakar Tundlam <dtundlam@nvidia.com>
Reviewed-on: http://git-master/r/382727
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Paul Walmsley <pwalmsley@nvidia.com>