platform: tegra21: add API to check bondout status
the bond-out registers are in CAR space, which indicate whether
a device is present or not. usually device tree should have
information about available HW resources, but pre-silicon
environment often is limited, and it's not desirable to define
a dts for each and every platform variants. therefore, reading
out bond-out register bits helps further confirm whether device
really exists or not.

Change-Id: I5e3504316a2c1b68a9c748ce97a71ba7b77ca3bb
Signed-off-by: Bo Yan <byan@nvidia.com>
Reviewed-on: http://git-master/r/385391