ARM: tegra132: fix CPU clock suspend/resume
A subset of the CPU clock registers are managed by Denver and
mapped at a separate aperture (0x70040000). For such registers,
we need to call clk_(readlx|writelx) which performs the access
to the appropriate aperture based on the SOC, given the driver
is shared by both T12x and T13x.

Bug 1488366

Change-Id: I344179c2626c94d144235814b843a1cb21b502d9
Signed-off-by: Peng Du <pdu@nvidia.com>
Reviewed-on: http://git-master/r/387363
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Aleksandr Frid <afrid@nvidia.com>
Reviewed-by: Adeel Raza <araza@nvidia.com>
Reviewed-by: Alexander Van Brunt <avanbrunt@nvidia.com>