arm: tegra: thermal: use correct zone pointer
The global index of thermal zones from thermal framework will be
different from the 0-based soctherm driver's own index. Use the global
index to notify framework, not the local index. Local index is used to
program the HW, and is derived from the stored devdata.

This fixes hw programming of thresholds in the wrong zone when written
from the trip_point_x_temp sysfs node.

This doesn't directly affect the bug mentioned below, but fixes a
problem that occurs when trying to repro it.

Bug 1482109

Change-Id: Ib03b719a51d296d3b77e99c53bcb68742576e3bb
Signed-off-by: Diwakar Tundlam <dtundlam@nvidia.com>
Reviewed-on: http://git-master/r/387984
Reviewed-by: Automatic_Commit_Validation_User