iio: core: add link file for iio_device
When iio device get registered, the iio device name has the
unique ID and all sysfs created under this device. The unique
id depends on the order it is called and so does nto give the
consistent deviceid for given device.

Hence add the link to point the iio device with unique name.

Change-Id: I0442de9fe27ba0c0e182b9469e5dc707dc31c850
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-on: http://git-master/r/391881
Reviewed-by: Timo Alho <talho@nvidia.com>