platform: tegra: pmc: read back thermtrip scratch regs to ensure they were set
Under certain conditions, the PMC can silently block writes into its
scratch registers.  This can be a big problem when the values being
written are intended to shut down the PMIC after a critical thermal
shutdown event.

In this patch, we read back the PMC scratch registers to ensure that the
values contained therein are the ones we programmed.  The code will
call WARN() if there is a mismatch.  It probably should just shut down the
device.

Bug 1482109

Change-Id: Idb6c17a78c2b512114dbb0b2ad6615875ecbe5c4
Signed-off-by: Paul Walmsley <pwalmsley@nvidia.com>
Cc: Diwakar Tundlam <dtundlam@nvidia.com>
Reviewed-on: http://git-master/r/392201
GVS: Gerrit_Virtual_Submit
Reviewed-by: Diwakar Tundlam <dtundlam@nvidia.com>
Tested-by: Diwakar Tundlam <dtundlam@nvidia.com>