ARM: tegra: vcm30t124: Added DTS file for V3A.
Initial created dts file for vcm30t124 Android.

A serial port for all ardbeg devices is uartc and it's
enumerated on ttyTHS2.
But it's uartd on vcm30t124 and uartd is enumerated on ttyTHS3.
So changed the ordering so that uartd can be enumerated on ttyTHS2.

bug 1440706

Change-Id: I6e226812402bc75991ece96c335b492f949f7976
Signed-off-by: Phoenix Jung <pjung@nvidia.com>
Reviewed-on: http://git-master/r/395661
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Sandeep Trasi <strasi@nvidia.com>