input: gpio-keys: get button status through gpio for irq based key
The gpio keys supports the key functionality through irqs. The driver
expect that key interrupt should be generated continuously till the
key is pressed.
When configuring with Palmas PMIC onkey, it is found that interrupt is
getting generated only once when it is pressed. There is no interrupt
if key is released. This way it is difficult to find that key pressed
duration is short or long. Provide the mechanism to find out the key
state through GPIO so that key press is detected through the interrupt
and state will be determined by the gpio.

This is based on
	http://git-master/r/375884 on bug 1392048
	by Vineel Kumar Reddy Kovvuri <vineelkumarr@nvidia.com>

Change-Id: I7ea9a98cc6e3d888143a41939a89daa34eae7f0a
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Signed-off-by: Vineel Kumar Reddy Kovvuri <vineelkumarr@nvidia.com>
Reviewed-on: http://git-master/r/396850