video: tegra: nvmap: Fix misleading comment
The comment would have led one to believe that the
NVMAP_PP_ZERO_MEM_FILL_MIN define would disable the page pools
entirely but that is not the case, instead it essentially
disables the background allocator. This is due to the fact that
without zero'ed pages memory is inserted right back into the
page pools instead of being freed to the system.

Change-Id: Id6e7664609f4a6d776febcfc9b9d751755806ac0
Signed-off-by: Alex Waterman <alexw@nvidia.com>