video: tegra: nvmap: increase set/ways threshold for denver
Doing full invalidates (and similarly flushes) causes everything
to be invalidated including the MTS translations. Thus all
translation must be redone/revalidated which is a very expensive
operation.

However, it still makes sense to do full invalidates and flushes
is the buffers are big enough. Thus we just increase the threshold
to 16 MB instead of the previous 2 MB.

Change-Id: I3648dd820051535cbc5e9390a0194b4e2041cff0
Signed-off-by: Alex Waterman <alexw@nvidia.com>
Reviewed-on: http://git-master/r/407858
Reviewed-by: Mandar Padmawar <mpadmawar@nvidia.com>
Tested-by: Mandar Padmawar <mpadmawar@nvidia.com>