video: tegra: host: Reorder job completion
Job completion currently first turns off the hardware and then
checks which jobs were completed . This is vulnerable as we
cannot turn off the hardware until jobs have been completed.

This ordering was introduced to avoid race between channel release
and devfreq callbacks. At the time the ordering was selected, it
worked correctly as we were not relying on runtime pm and therefore
we did not start power down sequence at that point.

However, with runtime pm we may start powering down the device at
this point. As we no longer have the race for scaling (it does not
use .deinit()/.init() callbacks) it is safe to reorder this sequence
back to the original form.

Bug 200002762

Change-Id: Ida6022b6878d2fa7868ac3369a95240f895fb2ac
Signed-off-by: Arto Merilainen <amerilainen@nvidia.com>
Reviewed-on: http://git-master/r/409598
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Terje Bergstrom <tbergstrom@nvidia.com>