video: tegra: host: Reduce timeout on close
When we close the application, we currently simply continue
processing work for the application that was already closed.
In most cases this approach is safe, however, in cases where
the work is dependent on user space this approach may make
the engine unusable for no reason.

This patch makes job timeout values to be shortened to 500ms
if the application closes the channel.

Change-Id: I143f6b442f80d5ba47defccdc2aada7667267ffb
Signed-off-by: Arto Merilainen <amerilainen@nvidia.com>
Reviewed-on: http://git-master/r/416244
Reviewed-by: Terje Bergstrom <tbergstrom@nvidia.com>
Tested-by: Terje Bergstrom <tbergstrom@nvidia.com>