video: tegra: host: Update sp value before incrs
Currently, finishing job increments assumes that the syncpoint value
has not changed during the timeout routine. However, this is done
just before stopping the channel and resetting the engine and therefore
there is possibility that the engine has finished the work just after
the check.

This patch adds an additional check to ensure that the code does not
accidentally do too many increments.

Bug 1517429

Change-Id: I4ea7c1635ff78f7d991bdf0bb35160a7bd5e186f
Signed-off-by: Arto Merilainen <amerilainen@nvidia.com>
Reviewed-on: http://git-master/r/417043
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Terje Bergstrom <tbergstrom@nvidia.com>