iommu/tegra: smmu: use iommus to get swgid from dt
If smmu wants to support various tegra-smmu as well as arm-smmu, a
common format is needed for representing the device specific properties
such as the iova range and mapping alignment,
number of prefetch pages required, if a gap page is required.

iommus property is the upstream way to specify these details per device.
If iommus property is present, construct swgid bitmap from the two
cells following the phandle. In case iommus property is not present
that means we have not upgraded, hence fall back to existing way.

Bug 1463797

Change-Id: Ic3cb148e1091e766d98182b3017035441abe826f
Signed-off-by: Sri Krishna chowdary <schowdary@nvidia.com>
Reviewed-on: http://git-master/r/412645
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Krishna Reddy <vdumpa@nvidia.com>