platform: tegra: nvadsp: Add msgq support
The message queue type has functions for adding and removing
messages in the order sent. Storage space for the message queue
and message buffer must be preallocated prior to initialization.
The queue depth is determined by the initializer up to a maximum
size of 8192 words to accommodate multiple, variable sized messages.
This is ported from ADSP side implementation of the same.

Also remove stale nvadsp_cmsg_queue.c.

Change-Id: I5ec23ca8970584f543ae92887928af508ae47f0d
Signed-off-by: Sumit Bhattacharya <sumitb@nvidia.com>
Reviewed-on: http://git-master/r/414173
Reviewed-by: Viraj Karandikar <vkarandikar@nvidia.com>
Reviewed-by: Sachin Nikam <snikam@nvidia.com>
Reviewed-by: Ajay Nandakumar M <anandakumarm@nvidia.com>
Reviewed-by: Nitin Kumbhar <nkumbhar@nvidia.com>