power: lc709203: add TEMP_INFO property support
Based on platform design, the fuel gauge lc709203 can sense the battery
temperature. On some design, it will nto be able to sense the battery
temp but on this case, the battery temp need to be monitor outside of this
device and need to update the device register for proper calcalutaion
of charger/capacity.

Add support for property  TEMP_INFO on which it will expose sysfs as temp
and provide the battery temp monitored from outside driver.

Change-Id: Ia29ab41d169d68a1d5447d3672b1bfb15ad8eea7
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-on: http://git-master/r/422605
Reviewed-by: Automatic_Commit_Validation_User