ARM: tegra: do not provide extcon name by default
If extcon client is in DT then it is not require to connect
client by extcon name and so it is not required to provide name.

The client can get connected with the handle of extcon driver.

Remove extcon names by default from PMIC DT node and only add this
names if client is in non-DT.

Change-Id: Iaf1707bef6514cdcc54b1af31d1b05396c6890da
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-on: http://git-master/r/428876