ARM: tegra: dvfs: Support Tegra21 DFLL idle override
On Tegra21 platforms DFLL control settings can be overridden in h/w
when CPU is entering certain idle states (CC3 and CC4). This commit
modified CL-DVFS driver to support override mechanism:

- added Tegra21 compatibility to CL-DVFS driver
- added TEGRA_CL_DVFS_HAS_IDLE_OVERRIDE flag, and specified it in SoC
match data for T210 SoC
- updated idle override settings to follow DFLL mode and output minimum
if the above flag is set
- changed entry to Open Loop mode when 1-wire PWM PMIC interface is
used, so that PWM is left enabled (in configuration when override is
supported, there is a separate idle controller that turns PWM on/off).

Bug 1529555

Change-Id: I47ea43851766c7717435c43423b1e45569c08f97
Signed-off-by: Alex Frid <afrid@nvidia.com>
Reviewed-on: http://git-master/r/432968
Reviewed-by: Yu-Huan Hsu <yhsu@nvidia.com>