video: tegra: host: Fix devfreq suspend/resume use
This far we have called devfreq suspend callback when system suspend
has been issued. This patch modifies the driver so that we call both
suspend and resume callbacks and these callbacks are called as part
of device suspend/resume sequences. As part of this process, this
patch removes .suspend_ndev() callback.

Change-Id: Ia32805ae33e60d8abec59e91230379e96d291637
Signed-off-by: Arto Merilainen <amerilainen@nvidia.com>
Reviewed-on: http://git-master/r/435534
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Deepak Nibade <dnibade@nvidia.com>
Reviewed-by: Shridhar Rasal <srasal@nvidia.com>