[WAR] staging: iio: light: iqs253: make sensor always on
Keep sensor always on for its proper functioning.
It seems, we do not work fine with sensor turning on when
needed and turning it off when not in use. This behavior
could be because of the time it takes to adjust to the
environment but the root cause is not yet clear.

Bug 1499925

Change-Id: I57a5949e01f8d9cf68e50d877a2dae6925592ffd
Signed-off-by: Sri Krishna chowdary <schowdary@nvidia.com>
Reviewed-on: http://git-master/r/436506
(cherry picked from commit 044f13382fb833eacc8cf93d8cb1bba729faba76)
Reviewed-on: http://git-master/r/436422
(cherry picked from commit db3c76570b1752dbfc4c1b5bcc0cbac5287c08c7)
Reviewed-on: http://git-master/r/448801
Reviewed-by: Sachin Nikam <snikam@nvidia.com>