ARM: tegra: dvfs: Don't protect dvfs data in early reasume
Don't protect dvfs data with mutex in late suspend/early resume.
DVFS operations are suspended at that time, but during syscore resume
it may be necessary to update DVFS data to changes happened while the
system was in suspend (in praticular memory configuration is chnaged
by boot-rom on suspend wake).

Change-Id: I8df278dc2e58e5ea1508d187afe0b8f4d8ccbe22
Signed-off-by: Alex Frid <afrid@nvidia.com>
Reviewed-on: http://git-master/r/438423
Reviewed-by: Gaurav Sarode <gsarode@nvidia.com>
Reviewed-by: Yu-Huan Hsu <yhsu@nvidia.com>