nohz: tick: retry for late sched_timer enqueue
When CPU is operating at low frequency and CPU leaves idle at a time
pretty close to jiffies boundary, hrtimer_start_expires() is likely
to end up with -ETIME. That means the desired expire time of
sched_timer is already in the past. In this case, jiffies update is
going to take place in softirq context which means it can be delayed
for arbitrary time. Kernel timekeeping has issue during this
timeframe.

This commit exames sched_timer after it gets enqueued. If it's a late
enqueueing, cancel the timer, update jiffes and try to enqueue again.

bug 200010487

Change-Id: Ic905020d36a8b5c897028729100783b851b80b3a
Signed-off-by: JC Kuo <jckuo@nvidia.com>
Reviewed-on: http://git-master/r/447635
GVS: Gerrit_Virtual_Submit
Reviewed-by: Sai Gurrappadi <sgurrappadi@nvidia.com>
Reviewed-by: Laxman Dewangan <ldewangan@nvidia.com>