staging: android: fiq_debugger: dump bt & regs
Override the debugger's FIQ handler to dump backtrace and regs for CPU0. We
have no way to get to the backtrace other than using the console for now. This
helps us get a backtrace from devices in the field.

The FIQ is set to be broadcast to all CPUs. Dump the context of each CPU
individually.

Bug 1511804

Change-Id: I0fb6699f6846a4ebb55fd1fc7c51dd9ce1fb1f20
Signed-off-by: Varun Wadekar <vwadekar@nvidia.com>
Reviewed-on: http://git-master/r/450798
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Kamal Balagopalan <kbalagopalan@nvidia.com>
Tested-by: Kamal Balagopalan <kbalagopalan@nvidia.com>
Reviewed-by: Mitch Luban <mluban@nvidia.com>