ARM: tegra: Notify cpufreq with actual set rate
Notify the cpufreq framework layer with the actual clock rate rather
than assuming the rate requested is the rate set. This could potentially
happen if the rate requested doesn't match the rate output by CL-DVFS
due to the granularity it supports.

Not doing so could potentially ruin the state of some cpufreq governors
due to the feedback loop.

Bug 1536224

Change-Id: I6d61b45689b9ae2c492d117b720f6dd106b300c2
Signed-off-by: Sai Gurrappadi <sgurrappadi@nvidia.com>
Reviewed-on: http://git-master/r/454162
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Diwakar Tundlam <dtundlam@nvidia.com>