nvmap: page pools: hide internal lock from nvmap_handle.c
The internal pool lock is exported so that nvmap_handle can lock
it, call a *_locked function, and then unlock it.  Provide a
version of the *_locked functions that takes the lock, remove
the lock and unlock helpers, and make the lock private to
the pools again.

Change-Id: I5a99753058e43161d50a0c61f3a984655cd7cd35
Signed-off-by: Colin Cross <ccross@android.com>
Signed-off-by: Krishna Reddy <vdumpa@nvidia.com>
Reviewed-on: http://git-master/r/664671
GVS: Gerrit_Virtual_Submit