irqchip: tegra: set affinity only if irq is enabled
We are using same register for affinity and interrupt
control. As such there is no register for affinity
setting in LIC. When any CPU is hot-unplugged this
function is called for interrupts causing enabling
all interrupts on CPU0.

Hence set affinity only if the interrupt is enabled
for any of the possible CPUs.

Change-Id: Ic8fff5ec72af99ea663d9d9a9d8c589959b58a0a
Signed-off-by: Prashant Gaikwad <pgaikwad@nvidia.com>
Reviewed-on: http://git-master/r/486912
Reviewed-by: Sivaram Nair <sivaramn@nvidia.com>
Tested-by: Sivaram Nair <sivaramn@nvidia.com>
Reviewed-by: Bo Yan <byan@nvidia.com>