video: tegra: host: Never release firmware
Firmwares are currently released at the time we close the channel
device node. This causes unexpected side effects in cases where
the finalize_poweron is called only once to get the device booted
during the first time the device is accessed.

This patch fixes the issue by not releasing the firmware after it
has once loaded.

Bug 200033244

Change-Id: I8fc6748846e56ac0cc347feb170b239533a47252
Signed-off-by: Arto Merilainen <amerilainen@nvidia.com>
Reviewed-on: http://git-master/r/495220