tegra: fb: Notify fbcon DC is unavailable
In the absence of hotplug, monitor data and mode database is removed
and DC is switched off. However, fbcon is unaware.
This patch adds the change to fb_info state to suspend while updating
monitor specs. This state will denote that the hardware is unavailable.
Info will be restored to state running once modelist is available.

Bug 200031817

Change-Id: I0d4a51fdaa383d9201ab299d602fb3cc9ed93f52
Signed-off-by: Pankaj Dabade <pdabade@nvidia.com>
Reviewed-on: http://git-master/r/496027
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Venkat Moganty <vmoganty@nvidia.com>