arm64: tegra: Disable Ethernet for Foster
We kind of pre-maturely enable Ethernet for Foster. None of the devices
currently have a MAC address programmed. It seems like Android crashes
on startup while trying to bring up the Ethernet in that case.

As a WAR, disable Ethernet for now. Need to figure out the right
strategy to enable this.

Change-Id: I842b2ee53f0e4910579eef2c1f65da6fcc0142e3
Signed-off-by: Aly Hirani <ahirani@nvidia.com>
Reviewed-on: http://git-master/r/498069
Reviewed-by: Raymond Poudrier <rapoudrier@nvidia.com>
Tested-by: Raymond Poudrier <rapoudrier@nvidia.com>