video: tegra: host: Move falcon firmware sw inits
This patch moves falcon firmware to be loaded as part of falcon
boot process. In order to keep consistency over drivers, move the
sw initialisation in the beginning of corresponding boot function
in all falcon units.

Also fix build warning of unused variable in case bootloader is not
used.

Bug 200006528

Change-Id: I63f39627f854af3883d07f8ac71063185271205a
Signed-off-by: Arto Merilainen <amerilainen@nvidia.com>
Reviewed-on: http://git-master/r/499787