arm64: tegra: pinmux/gpio files for Foster-E
Foster is significantly different enough in HW from Loki to now need a
different pinmux and gpio files.

This copies the Loki pinmux/gpio files and updates the pins that are
different between the two. Also moves the uart over audio section into
the Loki file instead of the gpio file

Bug 1552619

Change-Id: I1bc6970d4d142243254a0d763afbc18798e3e1ea
Signed-off-by: Aly Hirani <ahirani@nvidia.com>
Reviewed-on: http://git-master/r/538668
Reviewed-by: Raymond Poudrier <rapoudrier@nvidia.com>
Tested-by: Raymond Poudrier <rapoudrier@nvidia.com>