video: tegra: host: free sync point synchronously
For a host managed sync point in nvhost_free_syncpt(),
we first check if MIN != MAX and in that case print an
error and return without freeing the sync point.

This causes a sync point leak.

To fix this, wait until MIN == MAX instead of returning.
We wait for MAX_TIMEOUT in this case. Even if sync point
is stuck, recovery process should help to free the
sync point

Bug 200039324

Change-Id: If5240e007e26ff4fd819ec47694dd56eeaa82f6a
Signed-off-by: Deepak Nibade <dnibade@nvidia.com>
Reviewed-on: http://git-master/r/542638
Reviewed-by: Arto Merilainen <amerilainen@nvidia.com>
Tested-by: Arto Merilainen <amerilainen@nvidia.com>