arm: tegra: Read throttle table with an offset
Since max cpu frequency depends on the chip sku the throttle table entries
also varies depending on the sku. Slower sku ones will start with smaller
frequencies.

This change gets the max cpu frequency info and according sets the starting
index in the table.

By virtue of this, we need not change the dt contents of the throttle table
depending on the chip sku. Hence removed pruning in cpu-balanced table.

Change-Id: Ia3b8c0bc22abb770728f5ac778546e7b965ebf2e
Signed-off-by: Arun Swain <arswain@nvidia.com>
Cc: Diwakar Tundlam <dtundlam@nvidia.com>
Reviewed-on: http://git-master/r/542895
Tested-by: Diwakar Tundlam <dtundlam@nvidia.com>
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Diwakar Tundlam <dtundlam@nvidia.com>