ARM64: tegra: t210: DT: remove CAM_RST pin
It is caused by a hw issue on the ERS board(E2220), the CAM_RST (ps4)
pin is routed to the pmic board(E2174) as the enable pin of
EN_VDD_CAM(1.8v of U10), but it is also routed to the camera board as
the reset pin of each camera device.

Our camera drivers needs a WAR not use it any more.

This issue affects both the e2249 camera board and the e2128 6-camera
board.

bug 1567112

Change-Id: Ib9a25aa57bde169497d3dbe98cc857e2dab395af
Signed-off-by: Charlie Huang <chahuang@nvidia.com>