edp: sysedp: support capping GPU by mW
Extend sysedp_dynamic_capping to support GPU power caps (in mW) as
an optional alternative to GPU frequency caps (in kHz)

Capping in mW has a notable limitation relative to capping in
kHz. When capping in kHz, the CPU vs GPU priority decision depends on
the relation between current GPU frequency and the
cpu-priority-cap-on-gpu-frequency. That logic is absent when capping
GPU in terms of mW.

Bug 1558347

Change-Id: I8e77ae48d35618d9dac3122f7493d790e4af7d78
Signed-off-by: Matt Longnecker <mlongnecker@nvidia.com>
Reviewed-on: http://git-master/r/590938
Reviewed-by: Timo Alho <talho@nvidia.com>
Reviewed-by: Steve Rogers <srogers@nvidia.com>
Reviewed-by: Bo Yan <byan@nvidia.com>