cpuquiet: Fix cpu/online nodes update
When cpuquiet is enabled, it hotplugs the secondary
CPU, but since the implementation for hot-plug in
drivers/base/core has changed, the new status is not
reflected in the respective cpuX online nodes.

This patch replaces the cpu_up/cpu_down calls with device_offline
and device_online thereby keeping the cpu/online nodes in sync.

Bug 200039095

Change-Id: I0aaaa169aa2f512387babce6fbb6ecd4a2d7d7af
Signed-off-by: Ishan Mittal <imittal@nvidia.com>
Reviewed-on: http://git-master/r/592103
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>
Tested-by: Bharat Nihalani <bnihalani@nvidia.com>