video: tegra: dc: Update next EMC freq computation
The display algorithm for determining the minimum necessary BW is to
start at a low EMC frequency and check. If the frequency is too slow
then the display adds 1 to that frequency and rounds. That was expected
to then round to the next highest frequency. However, actually, there
is a small window in which the clock frequency is rounded down.

This patch simply adds 1000000 Hz instead to avoid that small window
where the frequency is rounded down.

Bug 1576403

Change-Id: I32902a2b82bba6b53465f4f4d4dbbd9ea01b383d
Signed-off-by: Alex Waterman <alexw@nvidia.com>
Reviewed-on: http://git-master/r/604303
Reviewed-by: Jon Mayo <jmayo@nvidia.com>
Tested-by: Jon Mayo <jmayo@nvidia.com>