host1x: Implement engine reset
Currently host1x cleans up the cdma and makes the additional
increments. However, the engine is not reset and as a result the
engine is likely unusable on the subsequent jobs.

This patch implements a callback to reset module after timeout. The
actual reset is then done in the driver code.

Bug 1704292

Change-Id: I360bb6d101be731c09b72cc96db86c4d86e46627
Signed-off-by: Arto Merilainen <amerilainen@nvidia.com>
Signed-off-by: Andrew Chew <achew@nvidia.com>
Signed-off-by: Shridhar Rasal <srasal@nvidia.com>
Reviewed-on: http://git-master/r/837823
Signed-off-by: Mikko Perttunen <mperttunen@nvidia.com>
(cherry picked from linux-3.18)