video: tegra: host: Rework job serialization
Currently job serialization is done in the beginning of each submit
which is ok as long as the submits are using the same syncpoints.

However, as we are going to support also a mode where multiple
channels can use the same engine and each of the channels will use
their own syncpoints. Therefore channel serialization must be done
at end of work instead of beginning of work.

This patch also enables forced op_done for the clients (VIC) using
serialization feature. This will ensure that we do not stop CDMA
while holding the MLOCK.

Bug 200041935

Change-Id: If13502cee4d24b009670baf6541ab53837491e45
Signed-off-by: Arto Merilainen <amerilainen@nvidia.com>
Reviewed-on: http://git-master/r/660248