input: misc: Add input type sensor driver for ST8.
Bug 200063068

Currently our sensor support the driver is iio
and HAL is MA5.20.
For to support ST8, can't upgrade the HAL to
MA5.20 from MA5.14.
Because the calibration data store format issue.
And HAL MA5.14 the kernel driver is input.
So we need to support both driver type.

Change-Id: I68f90f47f2dbe0e0817feec7e5142a4ea6d38a89
Signed-off-by: Steven Kao <skao@nvidia.com>
Reviewed-on: http://git-master/r/663061
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Robert Collins <rcollins@nvidia.com>