pstore: rtrace: support register physical address
The virtual addresses are close to impossible to decipher. Save
register physical addresses instead. If virtual to physical
translation is failing, save virtual address instead.

In addition, save the value written to register.

This is done for aarch64 only. A reference implementation is provided
for aarch32, but it has not been verified, so kernel build for ARMv7a
tegra chips still save virtual addresses.

Change-Id: Icca8063a1d89028ffff251f8a0689fef73ac617c
Signed-off-by: Bo Yan <byan@nvidia.com>
Reviewed-on: http://git-master/r/663995
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit