arm: dtb: st8: change compatible for sensor driver
Bug 200063068

Currently our sensor support the driver is iio
and HAL is MA5.20.
For to support ST8, can't upgrade the HAL to
MA5.20 from MA5.14.
Because the calibration data store format issue.
And HAL MA5.14 the kernel driver is input.
So we need to support both driver type.

Change-Id: I57328822e1750ef61872ae7338a32542e3423f38
Signed-off-by: Steven Kao <skao@nvidia.com>
Reviewed-on: http://git-master/r/664201
GVS: Gerrit_Virtual_Submit
Reviewed-by: Robert Collins <rcollins@nvidia.com>