tegra: ivc: add support for channel reset
Allows re-initialization of IVC channels from an unknown state to the
initial state. Services using this feature are required to call
tegra_ivc_channel_reset() before making any other IVC calls, and must
call tegra_ivc_channel_notified() in response to all IVC notifications.

For backwards compatibility, the all-zero state is initialized and ready
for communication, so existing services that never use the new calls
will still work, but are not guaranteed to recover from a non-graceful
restart.

This change deprecates the ops parameter to tegra_hv_ivc_reserve(),
because callbacks are not supported for services that handle reset,
which will be required in the future.

Implement reset for the IVC device file interface, which requires the
device IRQ handler to call tegra_ivc_channel_notified(). Switch this
handler to a threaded IRQ handler because nothing needs to be done at
IRQ level, and to avoid the need for disabling IRQs when reading or
writing to synchronize with reset handling.

Drop the loopback handler instead of adapting it to channel reset,
because we have no use for it. This also allows us to simplify the
reservation logic.

Bug 1567571

Change-Id: I24177df1f6b7dd458a853c09e6db3563ade1b0d1
Signed-off-by: Peter Newman <pnewman@nvidia.com>
Reviewed-on: http://git-master/r/665298
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Bahadir Balban <bbalban@nvidia.com>
Tested-by: Bahadir Balban <bbalban@nvidia.com>
Reviewed-by: Alexander Van Brunt <avanbrunt@nvidia.com>