iommu/tegra-smmu: fix unmap_pages() w/ 4KB and 4MB mix
unmap_pages() was broken when 4KB and 4MB pages are passed together. It
continued to unmap 4KB pages and ignored 4MB ones once it started. Now
unmap_pages() returns to a caller if it finds a 4MB page at PDE for the
next iteration to handle 4MB page in unmap_largepage().

The influence of this bug was, after incorrect unmap_pages(), the IOVA
was able to be allocated for a non-4MB mappable request. During this,
SMMU driver(ex: map_sg()) ignored PDE_NEXT and try to use the PDE as
page table. Depending on the data PDE has locate_pte() can end with
invalid PTE pointer, which caused a kernel crash at *pte.

Bug 1583295

Change-Id: I3f8e6e14d377cd01cadee803c387ab43e6cb9965
Signed-off-by: Hiroshi Doyu <hdoyu@nvidia.com>
Reviewed-on: http://git-master/r/665610
GVS: Gerrit_Virtual_Submit
Reviewed-by: Sri Krishna Chowdary <schowdary@nvidia.com>