spi: tegra: add support to have delay between packets
Add support for delay between packets. SPI controller does
not support this feature directly but it supports the
chipselct inactive between packets and the number of clock
cycle for inactive state.

So to keep chipselect active between packets with delay, chipselect
is converted to GPIO mode and set for proper level and configure
controller to CS inactive between packets to get the packet delay.

bug 1564116

Change-Id: Ic58b7d182678e6790bbed3335772481c8f636796
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-on: http://git-master/r/667672
Reviewed-by: Justin Kim (SW-TEGRA) <juskim@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Arun Shamanna Lakshmi <aruns@nvidia.com>