pwm: tegra: divisor and multiplier based on pretty-good algorithm
Add support to get the best divisor and multiplier based on pretty good
algorithms for 10 to 30KHz of frequency range. This offers less than 0.6%
of error on target frequency.

This method required to change the source clock frequency also and so if
it is enabled then only one pwm client will be possible as all pwm uses
the single clock source. Necessary check is added.

This feature is selected through DT and disable by default as multiple
user is most common configuration.

bug 1480402

Change-Id: I3b6474d70d556ab3aa99f1ba9226734307ec3ef6
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-on: http://git-master/r/667907
GVS: Gerrit_Virtual_Submit