misc: tegra-fuse: make chipid query overridable
The chip_id register has been moved in newer SOCs, however
the old driver (tegra-fuse) cannot be easily adapted since
it has code to address a number of quirks of legacy chips,
which makes the overhauling the old driver very error prone.
This change is the first step towards cleaning up the chipid
driver. It makes the query of chipid function overridable,
which is sufficient to address the chipid HW change. Once
the legacy chips branch out of mainline, we can revamp the
chip_id driver properly.

Change-Id: I5f7a31cc0ba4de83796f8ad9a5f613bedaf9d458
Signed-off-by: Peng Du <pdu@nvidia.com>
Reviewed-on: http://git-master/r/667978
Reviewed-by: Adeel Raza <araza@nvidia.com>
Tested-by: Adeel Raza <araza@nvidia.com>
Reviewed-by: Terje Bergstrom <tbergstrom@nvidia.com>