r8152: Disable random mac address
We want to explicitly hard fail if we've never fused a mac address. If
the mac address is not fused, there is a chance that the stability bits
for ethernet is also not fused.

In that scenario, with the random mac address, we'll get a user who
thinks his ethernet is fine (because of the random address), but
ethernet will quickly fail within minutes.

Bug 1611325
Bug 200220865

Signed-off-by: Aly Hirani <ahirani@nvidia.com>
Reviewed-on: http://git-master/r/707984
(cherry picked from commit 4860367055ac347a7c75540b4cc5f379ff99a8ce)
Reviewed-on: http://git-master/r/716077

Change-Id: I12f72b7e3406119468e13227268c2f863a48362a
Signed-off-by: Shreshtha SAHU <ssahu@nvidia.com>
Reviewed-on: http://git-master/r/1207723
GVS: Gerrit_Virtual_Submit
Reviewed-by: Yu-Huan Hsu <yhsu@nvidia.com>
Reviewed-by: Aly Hirani <ahirani@nvidia.com>