tegra: soctherm: add I/F to invalidate GPU tsosc
Add an interface for clients outside soctherm to be able to invalidate
GPU TSOSC.

The current list of clients include the PG driver.

NOTE: This change is needed for T210 based platforms. Things are different on
T18x platforms, but this code will not be used there, instead the BPMP Firmware
will take care of this situation.

Bug 1617400

Change-Id: Ifacb3cb635f894c0a8f1d79ebd697c6016bbf2c3
Signed-off-by: navneet kumar <navneetk@nvidia.com>
Reviewed-on: http://git-master/r/713449
Reviewed-by: Bo Yan <byan@nvidia.com>
Tested-by: Bo Yan <byan@nvidia.com>