clock: tegra21: Add lookup field for emc clock
Add lookup field for emc clock to avoid the "wildcard"
EMC clock matching.

This patch is part of a series of patches to fix a camera underflow
due to missing emc clocks in IsoMgr. To correctly fix that bug
we need to remove the EMC wildcard match. As a result all
usages of the EMC wildcard matching need to be removed.

Bug 200083696

Change-Id: I2c3238ed89507fcab31721ab9c02abddc41c64d9
Signed-off-by: Hoang Pham <hopham@nvidia.com>
Reviewed-on: http://git-master/r/712770
Reviewed-by: Bo Yan <byan@nvidia.com>