configs: tegra: unset CONFIG_LOCALVERSION_AUTO
Unset CONFIG_LOCALVERSION_AUTO in defconfigs, since it causes the build
to set the current git sha1 to the kernel version string. This appended
version confuses our magic scripts and affects build paths and kernel
module versions when the kernel is changed even slightly. The developer
usually knows which version he is using anyway, so the shortened sha1 is
unnecessary information.

Bug 200082598

Change-Id: I12cf67bc5042c848bdd295ade61ac960dfda3ead
Signed-off-by: Konsta Holtta <kholtta@nvidia.com>
Reviewed-on: http://git-master/r/713184
Reviewed-by: Terje Bergstrom <tbergstrom@nvidia.com>
Tested-by: Terje Bergstrom <tbergstrom@nvidia.com>