arm: tegra: vcm30t124: Re-setting clock frequency in vcm30t124 dtsi file
Since the qual team may change the default frequency for
controllers (or certain instances of it) it needs to be set
in top level dtsi of the board. Even if the default
frequency is set in tegra124-soc-base.dtsi it needs be
again set in the top level dtsi of board/VCM file for
clarity purpose.

Bug 200075505

Change-Id: Ia18f7ebd4d0b14cc60bc61be15c4a700335be0c0
Signed-off-by: Raveesh Kote <rkote@nvidia.com>
Reviewed-on: http://git-master/r/715192
Reviewed-by: Sandeep Trasi <strasi@nvidia.com>