video: tegra: host: Rework syncpt and ch policies
Currently we have two global flags for marking the syncpoint and
channel policies that should be used. However, we have recently
found out that it is necessary to support both policies and as a
result we introduced forced_map_at_open flag. This was solely
a WAR for the issue that the flags were global.

This patch..:
- Reworks channel allocation code to be unaware of the channel policy
- Reworks bus_client to deliver cookie always. Depending on policy
we use either instance specific or device specific cookies
- Merges syncpoint and channel policy flags into device specific
"resource_policy" flag.

Overall, this patch does not introduce a functional difference - it
only refactors the code into more maintainable form.

Bug 1614168

Change-Id: I5ef4d773aff89cb51e34f8bfe57b957546227078
Signed-off-by: Arto Merilainen <amerilainen@nvidia.com>
Reviewed-on: http://git-master/r/717722