video: tegra: host: Rework nvhost_vm hw ops
In our original plan nvhost hardware backing would have been entirely
responsible for hardware level management, however, this turned out
not to be the case. In contrast, the differences are abstracted
entirely behind DMA API. By noting this we can manage simply with
init and deinit functions which alter the vm structure data as
appropriate.

This patch refactors nvhost_vm hardware operations to base operations
entirely on init deinit functions. Initialization is postponed until
the first user requests a hardware address on the VM and
deinitialization is initiated when the last buffer is unmapped. As
the hardware operations maintain information of the backing, this
patch also simplifies static mappings by simply calling directly
hardware backing with the given parameters.

In addition, this patch allows disabling hardware backing on some
devices.

Bug 1243060

Change-Id: If8a982869c80e383dfa80a88704f40a581e6ee63
Signed-off-by: Arto Merilainen <amerilainen@nvidia.com>
Reviewed-on: http://git-master/r/720412
Reviewed-by: Automatic_Commit_Validation_User