arm64: tegra: Enable the INA226 on Jetson-E
On Jetson-E, there is an INA226 that is stuffed only on the FFPB
but not on the FFD (to measure DDR power). Since we currently use the
same dts file between the two, this change keeps the INA226 disabled
and enables it using the shim.

Bug 200038547

Change-Id: Ibd942a3b917152911ffcd51b0f3302e0784c2d3b
Signed-off-by: Jiukai Ma <jiukaim@nvidia.com>
Reviewed-on: http://git-master/r/723322
(cherry picked from commit 8776956804479e4ef8cff6b029f67d03986b5d64)
Reviewed-on: http://git-master/r/718855
Reviewed-by: Jay Bhukhanwala <jbhukhanwala@nvidia.com>
Reviewed-by: Laxman Dewangan <ldewangan@nvidia.com>