dvfs: tegra: Separate disabled-cdev/no-cdev
Handle differently cases, when
a) DVFS cooling device is present in DT with disabled status
b) No DVFS cooling device is present in DT or compatibility test failed

In case (a) don't generate error, and invalidate cooling device entry
for the respective DVFS rail to avoid further warnings. In case (b)
continue to generate the same errors and warnings as before this commit.

Bug 200004664

Change-Id: Iea5097dda1f54ead18828124a1d3d522c25f9f3d
Signed-off-by: Alex Frid <afrid@nvidia.com>
Reviewed-on: http://git-master/r/727333
Reviewed-by: Bo Yan <byan@nvidia.com>