arm64: initialize cpu hwcap for errata patching
For cluster-switched systems, not all CPUs may be online at the
time that CPU errata workarounds are applied. This will cause
dynamic errata detection based on CPU ID to miss some of the CPUs
on the system.

To work around this failure, this change allows the full set
of required errata workarounds to be specified in device tree data,
in smp_init_cpus, device node property "errata_hwcap" of all CPUs
are OR'ed together, then used to set cpu_hwcaps, this will ensure
the still offlined CPUs can also receive their errata workarounds.

If the device node property "errata_hwcap" is not defined, it will
have no effect on the dynamic detection based on CPU ID.

Change-Id: I8336b34aeece3f7409e3840dbd2d95abb256efb0
Signed-off-by: Bo Yan <byan@nvidia.com>
Reviewed-on: http://git-master/r/727120
Reviewed-by: Paul Walmsley <pwalmsley@nvidia.com>