ARM: tegra: hawkeye: set SD2 to FPS on shutdown
SD2 rail is always ON and can not be turn off when system is
alive. Also this is required to be OFF when PMIC is in OFF state.

Configure this rail to FPS1 in shutdown to turn it off in PMIC
OFF state.

bug 1624701

Change-Id: I5aee44e5c8d755af6312f07a99bca4ce94b9d128
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
(cherry picked from commit 75494e6be5f4260bdf488d965e8056c888c0f85c)
Reviewed-on: http://git-master/r/730143