platform: tegra: emc: Use regular timer
Use a regular timer (as opposed to a deferred timer) for the EMC
periodic calibration timer. It is possible that the deferred
timer is deferred for long enough that the clock-tree drift in
the DRAM is large enough to cause instability. For example,
suppose a game is played and warms up the DRAM. The periodic
training compensates as such. Then the game is stopped and a
few seconds later enters some sort of idle or suspend mode before
the EMC clock has fallen. When the device wakes up it will still
have the hot DRAM trimmer settings.

That said this is unlikely to occur in practice because normally
the EMC clock speed will dramatically fall off once high usage
has ended. However, some devices only have a few available freqs
and might not see a quick drop in frequency (perhaps the display
has a floor above the low freq which requires rounding to the
very high freq).

Bug 1633891

Change-Id: I52189250011a0651de56449e050b09a813ffc5a7
Signed-off-by: Alex Waterman <alexw@nvidia.com>
Reviewed-on: http://git-master/r/732769
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Aleksandr Frid <afrid@nvidia.com>
Reviewed-by: Yu-Huan Hsu <yhsu@nvidia.com>