drivers: tegra: Guard for cpufreq policy == NULL
The policy object for a CPU could be NULL due to a number of reasons one
of which is cpufreq hasn't initialized it yet in cpufreq_add_dev(). This
change avoids using policy if it is null.

Also fixed up the transition notifiers in tegra_update_cpu_speed() from
firing cpufreq_transition_end() incorrectly.

Bug 200094937

Change-Id: I24ae32c5d2729bd754ef4e6c94746979cbf95ee5
Signed-off-by: Sai Gurrappadi <sgurrappadi@nvidia.com>
Reviewed-on: http://git-master/r/735103
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Pritesh Raithatha <praithatha@nvidia.com>
Tested-by: Pritesh Raithatha <praithatha@nvidia.com>
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>