gpu. host1x: Pass register value in firewall
In gr2d and gr3d units the register offset was sufficient for determining
if the register in interest is used for storing a register value.

However, in VIC this is not the case. The operations are passed through
two registers, METHOD0 and METHOD1. Depending on content of METHOD0,
METHOD1 can be either address or data field.

This patch updates the firewall interface to deliver also the register
value to allow book-keeping inside the engine driver.

Bug 1698151

Change-Id: Iadab81dc98233d06a9cbfe433444c974b2dd9cf6
Signed-off-by: Arto Merilainen <amerilainen@nvidia.com>
Reviewed-on: http://git-master/r/823641
Signed-off-by: Mikko Perttunen <mperttunen@nvidia.com>
(cherry picked from linux-3.18)