ARM: tegra21: ERS: rails LDO1 and LDO8
Make LDO1 and LDO8 always ON with LP0 OFF.

Make these rails to be OFF in LP0 by connecting these rails with FPS1.

These rails are necessary for UPHY PAD and PLL. UPHY needs to be in
operating state in LP1. We can only turn these rails OFF in LP0.

Bug 200068549
Bug 1653521

Change-Id: If08dd287ec466bc36e8db0f5b13fe1c81f29be6a
Signed-off-by: TW Chiu <twchiu@nvidia.com>
Reviewed-on: http://git-master/r/756449
Reviewed-by: Ashutosh Jha <ajha@nvidia.com>
Tested-by: Ashutosh Jha <ajha@nvidia.com>
Reviewed-on: http://git-master/r/810580
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>