drm/tegra: Add falcon helper library
Add a set of falcon helper routines for use by the tegradrm client drivers
of the various falcon-based engines.

The falcon is a microcontroller that acts as a frontend for the rest of a
particular Tegra engine.  In order to properly utilize these engines, the
frontend must be booted before pushing any commands.

Bug 1698151

Change-Id: I87242f06d3dde8108454e24913f18c019cd5d6cf
Signed-off-by: Andrew Chew <achew@nvidia.com>
Signed-off-by: Arto Merilainen <amerilainen@nvidia.com>
Reviewed-on: http://git-master/r/823644
Signed-off-by: Mikko Perttunen <mperttunen@nvidia.com>
(cherry picked from linux-3.18)