gpu: host1x: Handle HOST1X registers directly
This moves handling of address registers in the HOST1X class directly
to the firewall code in host1x's job.c, so that individual clients don't
have to replicate this code. The list of address registers detected also
change from INDCTRL which is not actually an address register to INDOFF
and INDOFF2 which are.

Bug 1698151

Change-Id: I7e9270f3001f9197bd4ad3aa501c31d7ef2abe15
Signed-off-by: Arto Merilainen <amerilainen@nvidia.com>
Reviewed-on: http://git-master/r/823642
Signed-off-by: Mikko Perttunen <mperttunen@nvidia.com>
(cherry picked from linux-3.18)