video: tegra: nvmap: avoid calling 'nvmap_handle_put' for freed handle
API: 'nvmap_handle_get' is showing the following:
“nvmap_handle_get: nvmap_sanity attempt to get a freed handle”
And after that kernel oops were seen.

To avoid it, reorder the logic and take a ref count on next handle
though nvmap_handle_get() before releasing the spin lock. If it fails,
continue to next valid handle. By this we can ensure that handle has
a ref count.

Bug 200119763

Change-Id: I8fdbcc26d057345862c08ac3998ac8b5e0f17314
Signed-off-by: Amit Sharma (SW-TEGRA) <amisharma@nvidia.com>
Reviewed-on: http://git-master/r/767142
Reviewed-by: Krishna Reddy <vdumpa@nvidia.com>