clk: add __clk_debugfs_add_file to accept "struct clk *"
A module which creates virtual clock and doesn't have
clk_hw registered, will not able to use "clk_debugfs_add_file".
e.g. it wants to expose debugfs node for showing rate to
userspace but not tied the clk with clk_hw.

It exposes __clk_debugfs_add_file so that it can
accept legacy interface which accepts struct clk *.

Bug 1659882

Change-Id: If2e93d40a9c5ac64570ef94f9ea74e72d2b140ac
Signed-off-by: Puneet Saxena <puneets@nvidia.com>
Reviewed-on: http://git-master/r/776134
Reviewed-by: Sachin Nikam <snikam@nvidia.com>
Tested-by: Sachin Nikam <snikam@nvidia.com>
Reviewed-on: http://git-master/r/1127201
Reviewed-by: Ishan Mittal <imittal@nvidia.com>
Tested-by: Ishan Mittal <imittal@nvidia.com>