regulator: pmic-otp: add pmic-otp regulator
Some of system, it is require to not change th PMIC OTP
but client needs regulator handle to properly work.
On such cases, provide the regulator handle of pmic-otp
regulator which does not change anything for PMIC but
provide all possible regulator ops to client so that
client can work properly with any hack on regulator
requirements on client code.

bug 200131582

Change-Id: I512945193d3e171c8295bdd039b5002b7f3937a5
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-on: http://git-master/r/783983
GVS: Gerrit_Virtual_Submit
Reviewed-on: http://git-master/r/1149825
Tested-by: Shawn Joo <sjoo@nvidia.com>
Reviewed-by: Joshua Cha <joshuac@nvidia.com>