fbcon: Warn if no matching mode can be set
var_to_display() should always be able to find a matching node because
the content of struct fb_var_screeninfo was derived from a mode in the
modelist associated with the struct fb_info.

However there are cases where this can happen: when a mode is converted
to a struct fb_var_screeninfo and back again, rounding errors can result
in the mode no longer matching the original. When this happens it causes
a crash further down the road in fbcon_switch() because it will try to
access a NULL mode. To make matters worse, this crash happens with the
console lock held, making it almost impossible to diagnose without going
in and disabling the lock.

Add a warning if this case is detected to give a hint about why things
might be going wrong.

Bug 200130894

Change-Id: I27bc0dca8d7b4d4d4dafb3d8829a5940a2f28995
Signed-off-by: Thierry Reding <treding@nvidia.com>
Signed-off-by: Naveen Kumar S <nkumars@nvidia.com>
Reviewed-on: http://git-master/r/795367
(cherry picked from commit 5257f309ee32e6389a3538be4852c913dca14adc)
Reviewed-on: http://git-master/r/927724
Reviewed-by: Bibek Basu <bbasu@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Winnie Hsu <whsu@nvidia.com>