sched: CPU capacity aware wakeup task placement
Load balancing doesn't handle assymetric CPUs very well in that higher
capacity CPUs are not used for 'high load' tasks. This is caused by
the balancer trying to do smp nice balancing.

To fix this, we attempt to select the right sched group to place the task
in on wakeup. The cpu with enough spare capacity to fit the task is then
selected as the wakep target.

For big.LITTLE systems, this translates to trying to put big tasks on
big cores and little tasks on little cores.

Bug 1665314

Change-Id: I9947877399704bff5d92e4bdf353944a0b5c3a86
Signed-off-by: Sai Gurrappadi <sgurrappadi@nvidia.com>