adma: tegra: Add dma_start_index DT property
+ Adding support for optional dma_start_index property that
specifies the index of first physical channel to be used.
Index counting starts from zero. This change is required for
dual linux virtualization case where second guest need to skip
initial few physical channels that are assigned to the first
guest. If this property is not specified in DT, then no
physical channels are skipped.

Bug 200115263

Change-Id: I11c56afa758cdce4c5e56ec8ba642cb9db92af2e
Signed-off-by: Paresh Anandathirtha <paresha@nvidia.com>
Reviewed-on: http://git-master/r/800567
Signed-off-by: Seema Khowala <seemaj@nvidia.com>
Reviewed-on: http://git-master/r/817030
Reviewed-by: Automatic_Commit_Validation_User
Tested-by: Mihir Joshi <mihirj@nvidia.com>
Reviewed-by: Gaurav Sarode <gsarode@nvidia.com>