arm64: clean up reporting of bad_mode messages
There were two problems with the reporting of bad_mode
messages.  The first was that the output between the
different CPUs were interleaved rendering the output
unreadable.  The second problem was that interrupts
were not masked during the process.  As soon as an
interrupt window was opened a pending interrupt would
be taken and the bad_mode was re-entered.  This caused
the output to scroll continuously until the watchdog
fired.

The first problem was to place appropriate spin locks
around sections of code.  The second problem was solved
by forcibly disabling interrupts while in the die()
function.

Change-Id: Ie0ce6eddf60bf788cd0bd012f921b26813d6e568
Signed-off-by: Guy Sotomayor <gsotomayor@nvidia.com>
Reviewed-on: http://git-master/r/807703
GVS: Gerrit_Virtual_Submit
Reviewed-by: Richard Wiley <rwiley@nvidia.com>
Reviewed-on: http://git-master/r/841854
Tested-by: Richard Wiley <rwiley@nvidia.com>
Reviewed-on: http://git-master/r/1125505
Reviewed-by: Alexander Van Brunt <avanbrunt@nvidia.com>
Tested-by: Alexander Van Brunt <avanbrunt@nvidia.com>