platform: tegra: bootrom: add support for power off commands
Add support for power off commands fro bootrom. The commands
are provided from DT as follows:
pmc {
	bootrom-commands {
		reset-commands {
		};
		power-off-commands {
		};
	};
};

If power off and reset commands are not available from child node
then properties and command in node bootrom-commands are used. This
is done to support backward compatibility.

The commands properties are same as earlier.

bug 200146310

Change-Id: I3a2102a20424c81310a7c96041884b24ce6e9e15
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-on: http://git-master/r/829082