clk: tegra: Add Tegra210 device tree binding
Add a header file that defines the clock numbers for Tegra210. It is
meant to be included by device trees so that they can refer to the
clocks by symbolic name instead of numeric value.

Also add the device tree binding documentation which is largely the
same as for earlier generations of Tegra.

Extracted from a larger patch by Rhyland Klein <rklein@nvidia.com>.

Change-Id: I5bbda1950a7884419a8b2d6fbe6bf5ddd7c8fa07
Signed-off-by: Thierry Reding <treding@nvidia.com>