usb: gadget: Fix handling of unrecognized OS descriptor requests
If composite_setup() receives a control request for an OS Feature
Descriptor that is not Extended Compatibility ID or Extended Properties,
it falls through to enqueueing a response of length 'value' bytes which,
because the request is not recognized, is -EOPNOTSUPP.  Fix it so that
we instead bail out of composite_setup() without enqueueing a response.
The UDC driver is then responsible for halting EP0.

This is backported from chromeos commit:
c50ae94f78708e05e8eb4774b55cffc133047d44

Bug 200153616

Change-Id: I363c53a35a45af5cb571370ae18d098c2f169de5
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/313224
Reviewed-by: Benson Leung <bleung@chromium.org>
Signed-off-by: Mark Kuo <mkuo@nvidia.com>
Reviewed-on: http://git-master/r/843397
Reviewed-on: http://git-master/r/1165722
GVS: Gerrit_Virtual_Submit
Reviewed-by: ChihMin Cheng <ccheng@nvidia.com>
Reviewed-by: Ashutosh Jha <ajha@nvidia.com>