usb: gadget: tegra_xudc: fix unhalt ep failed
As EP_HALT register description, EP_HALT is valid only when the
Device Mode is in Run Mode. In provious code, ep_unhalt() is
called when device is in Disabled state. So in call ep_unhalt(),
EP status is wrong. This commit postpone to call ep_unhalt() to
when device mode change to Run Mode.

Bug 200154898

Signed-off-by: George Zhou <gzhou@nvidia.com>
Reviewed-on: http://git-master/r/921296

Change-Id: I222d066ca06915f8ad3a38af7a528ae41329c97f
Signed-off-by: JC Kuo <jckuo@nvidia.com>
Reviewed-on: http://git-master/r/1122436
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>