gadget: xudc: Don't poll for port reset completion
Quickcharge chargers are generating Port Reset Change (PRC) events upon
connection which never complete.  Since the XUDC appears to generate
PRC events on both 1-to-0 and 0-to-1 transitions of the Port Reset (PR)
bit, we can avoid polling for completion of the port reset by ignoring
the 0-to-1 transition.  The port reset will then be handled on the PRC
event generated by the 1-to-0 transition.  Do the same for WRC/WPR to
avoid a similar situation there.

This is backported from chromeos commit:
54a5a11c20f9d57c3c8912f01572715570ab6b2f

Bug 1706834

Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Signed-off-by: Mark Kuo <mkuo@nvidia.com>
Reviewed-on: http://git-master/r/837110

Change-Id: I8c0a58a9778b692ddff2771480223a843add8eb9
Signed-off-by: JC Kuo <jckuo@nvidia.com>
Reviewed-on: http://git-master/r/1122429
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>