usb: gadget: xudc: set PORTPM L1S accordingly
PortPM L1S bit is used to accept or rejest L1 entry request. When LPM is
disabled, PortPM L1S should be set to NYET, and when LPM is enabled, we
should also set the L1S to ACCEPT.

This is optional for now since lpm_enable value cannot be changed at
run time, but should be done if lpm_enable is configurable at run time
in the future.

This change is for the new XUDC driver (tegra_xudc.c).

Bug 1699527

Signed-off-by: George Zhou <gzhou@nvidia.com>
Signed-off-by: Mark Kuo <mkuo@nvidia.com>
Reviewed-on: http://git-master/r/837114

Change-Id: Ibc780a8019701c646aaf0e2c08a77943706ef840
Signed-off-by: JC Kuo <jckuo@nvidia.com>
Reviewed-on: http://git-master/r/1122433
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>