usb: gadget: xudc: correctly re-enable EP
When enabling an EP, we missed to check if the EP is already enabled.
When re-enabling the same EP, the "nr_enabled_eps" count will be
incorrectly increased by 1. This has caused unbalanced nr_enabled_eps
which is used as an indicator to switch from Configured state to
Addressed state when disabling endpoints in __tegra_xudc_ep_disable().
The result will be that XUDC stays in Configured state even if it should
be in the Address state.

Fix this by checking EP status first and if it's already enabled,
first disable the EP before enabling it.

Bug 200154992

Signed-off-by: Mark Kuo <mkuo@nvidia.com>
Reviewed-on: http://git-master/r/837112

Change-Id: I7820046156db9ed873e777b80295522cb18553e9
Signed-off-by: JC Kuo <jckuo@nvidia.com>
Reviewed-on: http://git-master/r/1122431
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>