video: tegra: nvmap: remove prot variable where unused
In do_cache_maint, handle is already checked whether it is cacheable
or not. If handle is cacheable, then nvmap_pgprot simple returns the
prot that is passed to it. So, avoid unnecessary call to it and also
clean up unnecessary prot variable.

In handle_page_alloc, prot is not used anywhere. So, remove it.

Bug 200140685

Change-Id: I6cd3413df353026f6847556731bbce7e27a53b11
Signed-off-by: Sri Krishna chowdary <schowdary@nvidia.com>
Reviewed-on: http://git-master/r/838461
GVS: Gerrit_Virtual_Submit
Reviewed-by: Pritesh Raithatha <praithatha@nvidia.com>
Reviewed-by: Amit Sharma (SW-TEGRA) <amisharma@nvidia.com>
Reviewed-by: Sachin Nikam <snikam@nvidia.com>