net:eqos: Register as an ISO BW manager client
Request a fixed Bandwidth of 80MBps, needed 60MBps
for write and 20MBps for read.

The "latency that can be tolerated by client in usec"
parameter passed to tegra_isomgr_reserve(), is set 0.
Which means "do not care" condition.

Since Automotive has fixed emc clock and mobile does
not use the EAVB use case, the latency value is
do not care for EQOS>

All these values are taken from the xls document
mentioned in the bug 1675616.

Bug 1675616

Change-Id: I73ebcce16439a08fbd56db172d9032186b935ea4
Signed-off-by: Virupax.Sadashivpetimath <virupaxs@nvidia.com>
Reviewed-on: http://git-master/r/933698
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: David Lim <dlim@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Sumeet Gupta <sumeetg@nvidia.com>