xhci-tegra: fix panic in tegra_xhci_set_host_mode
When firmware fails to start (clear CNR), driver will return the
allocated hcd to kernel and set tegra->hcd with NULL as well.
In this case, if there comes a extcon event, kernel panics in
tegra_xhci_set_host_mode() when driver tries to dereference from
tegra->hcd. This commit fixes it.

bug 200160468

Signed-off-by: JC Kuo <jckuo@nvidia.com>
Reviewed-on: http://git-master/r/921792

Change-Id: Ieeac5962edf64ca9c27d57c7889afeefe5c22ee1
Signed-off-by: JC Kuo <jckuo@nvidia.com>
Reviewed-on: http://git-master/r/1122460
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>