PM/Domains: Add/Modify as per upstream code
Modify code in domain.c so as to make it upstream compatible.
Currently we are retaining legacy code as well, because we
are in transition from legacy to upstream implementation. Once
we are in fully sync with upstream, we will remove legacy code.

Bug 200127732

Conflicts:
	drivers/base/power/domain.c

Change-Id: Id170293ebff018cf9b981f8a39cc19cfbe3d5d46
Signed-off-by: Sumit Singh <sumsingh@nvidia.com>
Reviewed-on: http://git-master/r/779148
(cherry picked from commit 925f30c9a39d136728dce57aa68a2cae900954fb)