net: wireless: bcmdhd: Avoid data pkt processing if credits is low
When running the EU wifi adaptivity test, ioctl timeouts are
observed as ctrl frames are out of bus credits and do not get sent
to the wifi chip. As a result, the OS resets the wifi chip to recover
the ioctl timeout. This results in EU wifi adaptivity test failure.

To prevent this from happening, avoid processing the data frames if
available credits is low.

Bug 1700549

Change-Id: Iaaf46112b1673088c27bfc5126b874dcdeb82b2d
Signed-off-by: Srinivas Ramachandran <srinivasra@nvidia.com>
Reviewed-on: http://git-master/r/999523
GVS: Gerrit_Virtual_Submit
Reviewed-by: Winnie Hsu <whsu@nvidia.com>