rtc: max77686: Fix unsupported year message
The max77686 RTC only supports a range of 99 years so instead of using
year 1900 as the base, the year 2000 is used. This means that 1900 to
1999 are unsupported years.

The driver was printing a warning for those values but was returning a
error so for consistency, print an error message instead and don't say
that a year 2000 is assumed, since the year is not set.

Also, it is better to use dev_* log functions instead of pr_* to print
information about the device in the kernel log in a standardized way.

This also allows to remove the local pr_fmt() defined macro.

Suggested-by: Krzysztof Kozlowski <k.kozlowski@samsung.com>
Signed-off-by: Javier Martinez Canillas <javier@osg.samsung.com>
Reviewed-by: Krzysztof Kozlowski <k.kozlowski@samsung.com>
Reviewed-by: Andi Shyti <andi.shyti@samsung.com>
Signed-off-by: Alexandre Belloni <alexandre.belloni@free-electrons.com>
(cherry picked from commit 1e5813bded3c6702841549ac2533680f410aea43)
Signed-off-by: Venkat Reddy Talla <vreddytalla@nvidia.com>

Change-Id: I0cf54589387ca3b16d8bc714130a9f7800eaf1c5
Reviewed-on: http://git-master/r/1115971
Tested-by: Venkata Jagadish <vjagadish@nvidia.com>
Reviewed-by: Venkata Jagadish <vjagadish@nvidia.com>