clk: Deprecate CLK_IS_ROOT
We don't use CLK_IS_ROOT but in a few places in the common clk
framework core. Let's replace those checks with a check for the
number of parents a clk has instead of the flag, freeing up one
flag for something else. We don't remove the flag yet so that
things keep building, but we'll remove it once all drivers have
removed their flag usage.

Change-Id: I54ade8dc9005c0e3506d704889a6cccb54cb5c9a
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>