usb: xhci: prepare ring correctly for ZLP transfer
When condition for ZLP (Zero Length Packet) is satisfied for Bulk OUT transfers,
an extra Zero Length packet is to be sent out after requested data is
transferred.

Currently, if the number of TRB slots available in the transfer ring
is exactly the same as the number of TRBs needed for the data packets, an
attempt will be made to equeue_trb when there is no space available. This
patch checks/prepares transfer ring correctly for the required number of TRBs
(data and ZLP) before attempting to enqueue them.

Also, when scatter-gather is involved, this patch ensures that ZLP is actually
queued.

Bug 1731030
Bug 1839120

Change-Id: Ib6fd12694f60f3c3409802e267d381f6821d725c
Signed-off-by: Haribabu Narayanan <hnarayanan@nvidia.com>
Reviewed-on: http://git-master/r/1022985
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Jui Chang Kuo <jckuo@nvidia.com>
Reviewed-by: Sumeet Gupta <sumeetg@nvidia.com>
Reviewed-on: http://git-master/r/1247892
GVS: Gerrit_Virtual_Submit
Reviewed-by: Ashutosh Jha <ajha@nvidia.com>