sched: Add hysteresis for capacity_aware_wake
Currently the CAPACITY_AWARE wake path tries to put a task on the
smallest capacity CPU it can find. It uses a 20% headroom to provision
future task growth but it doesn't use any hysteresis to provision for
any temporary dips in load.

This can cause unbounded thrashing if the load of a task oscillates
about the 20% headroom point. Fix this by introducing an additional 20%
guardband on the way down i.e (big->little) to provision for variablilty
in task load.

Change-Id: If1a43fd70521cefae51919c7ecd67e72d3a48f96
Signed-off-by: Sai Gurrappadi <sgurrappadi@nvidia.com>
Reviewed-on: http://git-master/r/1145099
Reviewed-by: Peter Boonstoppel <pboonstoppel@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>