clk: expose of_clk_get_from_provider()
Exposing this function makes the task of quering the clock
framework for a clock associated with an index a lot more
straight forward.
The alternative would be to create some sort of device-tree
fragment to be able to utilize the already exposed
of_clk_get_*() functions.

This kind of unusual requirement is used by the bpmp_clk test,
which need to aqcuire struct clk pointers to a wide set of
clocks in the system.

Change-Id: Iddb9ad06a7f642f073a619b37b0ac5ca077126fe
Signed-off-by: Stefan Kristiansson <stefank@nvidia.com>
Reviewed-on: http://git-master/r/1140393
(cherry picked from commit cf25375b7d366355d6721e715ff863babea3f8e5)
Reviewed-on: http://git-master/r/1178275
GVS: Gerrit_Virtual_Submit
Reviewed-by: Antti Miettinen <amiettinen@nvidia.com>
Tested-by: Antti Miettinen <amiettinen@nvidia.com>
Reviewed-by: Timo Alho <talho@nvidia.com>